refactor: Separate code from App.js to SearchPage.js
Took the source code from App.js that functions as a search tool
to ./components/pages/SearchPage.js.

In App.js, the exact path to route to the search page is '/search'.
This is based off of the base code (BooksAPI.js)
provided by Udacity so we must follow it as is.